fix(cpp): 修复Windows MSVC构建问题 (#258)
- 将编译选项从全局 add_compile_options 改为 target_compile_options
  解决 D9025 警告（/W3 与 /W4 冲突），避免影响外部依赖库
- 为 Windows MSVC 预创建多配置目录（Debug/Release等）
  解决 DLL 复制到不存在的 Release/ 目录失败的问题
- 添加 static_cast<int>() 显式类型转换
  解决 C4267 警告（size_t 到 int 的隐式转换）

fix(cpp): 修复Windows MSVC构建问题

- CI 使用 Visual Studio 生成器替代 Ninja，消除 D9025 警告
- 为 MSVC 预创建多配置目录，防止 DLL 复制失败
- 为 alibabacloud_credentials 添加 CMAKE_SKIP_INSTALL_RULES

+++ cpp/src/Utils.cpp
@@ -355,9 +355,9 @@ string Utils::getROASignature(const string &stringToSign, const string &secret) 
   unsigned char hashData[SHA_DIGEST_LENGTH];
   unsigned int hashLen;
   
-  HMAC(EVP_sha1(), secret.c_str(), secret.length(),
+  HMAC(EVP_sha1(), secret.c_str(), static_cast<int>(secret.length()),
        reinterpret_cast<const unsigned char*>(stringToSign.c_str()),
-       stringToSign.length(), hashData, &hashLen);
+       static_cast<int>(stringToSign.length()), hashData, &hashLen);
   
   Darabonba::Bytes bytes;
   bytes.assign(hashData, hashData + hashLen);
@@ -467,9 +467,9 @@ string Utils::getRPCSignature(const map<string, string> &signedParams, const str
   unsigned char hashData[SHA_DIGEST_LENGTH];
   unsigned int hashLen;
   
-  HMAC(EVP_sha1(), key.c_str(), key.length(),
+  HMAC(EVP_sha1(), key.c_str(), static_cast<int>(key.length()),
        reinterpret_cast<const unsigned char*>(stringToSign.c_str()),
-       stringToSign.length(), hashData, &hashLen);
+       static_cast<int>(stringToSign.length()), hashData, &hashLen);
   
   Darabonba::Bytes bytes;
   bytes.assign(hashData, hashData + hashLen);
@@ -743,9 +743,9 @@ string Utils::getAuthorization(const Darabonba::Http::Request &request, const st
   if (signatureAlgorithm == "ACS3-HMAC-SHA256") {
     unsigned char hashData[SHA256_DIGEST_LENGTH];
     unsigned int hashLen;
-    HMAC(EVP_sha256(), accessKeySecret.c_str(), accessKeySecret.length(),
+    HMAC(EVP_sha256(), accessKeySecret.c_str(), static_cast<int>(accessKeySecret.length()),
          reinterpret_cast<const unsigned char*>(stringToSign.c_str()),
-         stringToSign.length(), hashData, &hashLen);
+         static_cast<int>(stringToSign.length()), hashData, &hashLen);
     signature.assign(hashData, hashData + hashLen);
   } else if (signatureAlgorithm == "ACS3-RSA-SHA256") {
     // RSA signature - would need private key handling
